Encryption support
* Encryption support
* Updated README
* Improved logging in many classes
* Added communicationState channel which shows the bridge connection
state to the Paradox system
* Changed the priority of parameter panelType. Now the parameter is with
higher priority, i.e. if you provide valid panelType autodiscovery of
panel type will not be executed, saving initial connection time.
* Renamed several classes and interfaces to reflect more properly their
usage
* Renamed methods to reflect their usage better
* Improved lifecycle of model objects in case of disposal of bridge
handler
* Split the contents of the packet into two classes - packet and inner
class header which holds the header parts.
* Added listener interface (IResponseReceiver) for response receiving,
i.e. when response is received it is returned to the original sender by
using this interface.
* Parsing and decrypting of packets payload is now done inside the
Packet object (before it was in the communicator)
* Added handling of wrong packet response, i.e. live event which we
don't parse right now because we don't know what it means and how to
parse it.
* Improved overall handling of asynchronous situations when data may not
have been received but update is triggered (NPE, etc).
* Added more default values to non-mandatory parameters. Classified
mandatory parameters as mandatory in bridge XML
* Implemented argument handler for main method which is used for testing
of model as java application.
* Some unit tests

Signed-off-by: Konstantin Polihronov <polychronov@gmail.com>